chore(ui): move generic stores into 'generic' dir
