+++ ui/src/store/generic/GenericKeyValueStore.ts
